requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
python-dotenv
httpx
itsdangerous
google-cloud-storage
python-multipart